started adding main page design to center for pediatric dentisty-la-habra
